requirements.txt
fastapi
uvicorn[standard]
python-binance
pandas
pandas-ta
psycopg2-binary
Jinja2
requests